Add file structure, HTML and CSS templates
Add the file structure template used for making the webpage
projects. This consists of an index.html and an styles
directory at root, and a style.css file at the styles folder.
Also add the basic webpage template code for the index.html
and style.css files.

The file structure template is added to organize the project.
The webpage template is used as starting point for the HTML
and CSS code employed in the project.