Add images and finish minimal working html code
Add four images to the project. Add a link to a fontawesome
stylesheet and a navbar with a list of links to the
sections. Add a welcome section with a heading, a projects
section with a grid of image and links to my other projects,
an about section with lorem ipsum text and a contact section
with links to my github, codepen and email.

The images are used in the portifolio grid to preview the
projects. The link to fontawesome is used to get the symbols
for the contact links. The remaining additions are
requirements of the project.