Add minimal working css code
Add navlist id, removed 6 base colors to add 5, set * margin
and padding to 0 and set base font-size to 55%. Add minimal
working style for the navbar and the welcome section.

The navlist id will be used to style the navbar. The base
color are used to create the page color scheme. The margin,
padding and font-size are used as base for the style. The
minimal working styles are added to comply with the project
requirements.