Change change base size  and improve navbar style
Change font base size to 100%, add body font size, line
height and color, removes text decoration of links and
set their color. Change navbar background color and
justify-content, add block display and padding to the
links and add background color when hovering. Add
a flex display to navlis and chage the colors of the
welcome section's background image.

The changes to font base size and the body properties
is to improve the webpage responsiveness. The changes
to link decoration are part of the design choice, as
well as the navbar and welcome section. At this stage,
the navbar code is done and it will not be touched
until media query implementation.