Add missing content and finish  styling for mobile
Add divs and more headings to sections, remake the projects
-grid items adding divs, change the lorem ipsum with a
paragraph about NeBeWu, change the contact links and
add a footer. Also, change how links react with hover,
finish styling the four sections and style the footer.

The change to lorem ipsum is made to reflect a more
personal and creative feeling to the page. The footer
is added to leave a signature for the page. All the
remaining changes are to help directly or indirectly
with the customization of the page.