Tablet portrait and landscape styling
Change body font-size to 2rem and align navbar link list to
the start. Change project section height to 100vh, set
projects-grid to span more columns with auto-fit, increase
project-image to 30vh and set project-tile to fit them.
Set about section height to 70vh, set contact section to
70vh, readjust contact section padding, increase contact-
link size to 4rem and increase contact links between space.

The body font-size and the contact-link size changes are
to improve readability on large screens. The reamaining
changes are to improve the aestetics of the page.